pyproject.toml
[tool.poetry]
name = "podcast_release"
version = "0.1.0"
description = ""
authors = ["eucariot <eucariot@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"
python-decouple = "^3.4"
psutil = "^5.8.0"
paramiko = "^2.7.2"
scp = "^0.13.4"
google-api-python-client = "^2.10.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
